Refactors TabbedGridFragmentActivity to be abstract
